+++ src/booking/entities/booking.entity/booking.entity.ts
@@ -1,7 +1,7 @@
 
 import { AccessCredentialEntity } from 'src/access-credential/entities/access-credential.entity/access-credential.entity';
 import { IncidentEntity } from 'src/incident/entities/incident.entity/incident.entity';
-import { Review } from 'src/review/entities/review.entity';
+import { Review } from 'src/review/entities/review.entity/review.entity';
 import { Space } from 'src/space/entities/space.entity/space.entity';
 import { User } from 'src/user/entities/user/user.entity';
 import { Entity, PrimaryGeneratedColumn, Column, ManyToOne, OneToMany, OneToOne, JoinColumn } from 'typeorm';
