+++ src/review/review.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Review } from './entities/review.entity';
+import { Review } from './entities/review.entity/review.entity';
 import { ReviewService } from './review.service';
 import { ReviewController } from './review.controller';
